http server as component
