+++ components/led_http_server/CMakeLists.txt
@@ -1,0 +1,2 @@
+idf_component_register(SRCS "led_http_server.c"
+                    INCLUDE_DIRS "include")
