+++ components/led_http_server/led_http_server.c
@@ -1,0 +1,23 @@
+#include <stdio.h>
+#include "led_http_server.h"
+
+static esp_err_t get_handler(httpd_req_t *req)
+{
+    char *response_message = "URI Response ...";
+    httpd_resp_send(req, response_message, HTTPD_RESP_USE_STRLEN);
+    return ESP_OK;
+}
+
+void start_http_server()
+{
+    httpd_config_t server_config = HTTPD_DEFAULT_CONFIG();
+    httpd_handle_t server_handle = NULL;
+    httpd_start(&server_handle, &server_config);
+
+    httpd_uri_t uri_get = {
+        .uri = "/",
+        .method = HTTP_GET,
+        .handler = get_handler,
+        .user_ctx = NULL};
+    httpd_register_uri_handler(server_handle, &uri_get);
+}
