+++ main/hello_world_main.c
@@ -3,30 +3,12 @@
 #include "wifi_connection_handler.h"
 #include "nvs_flash.h" //non volatile storage
 #include "mdns_service.h"
+#include "led_http_server.h"
+
+
 
 
 static const char *TAG = "main";
-
-static esp_err_t get_handler(httpd_req_t *req)
-{
-    char *response_message = "URI Response ...";
-    httpd_resp_send(req, response_message, HTTPD_RESP_USE_STRLEN);
-    return ESP_OK;
-}
-
-void start_http_server()
-{
-    httpd_config_t server_config = HTTPD_DEFAULT_CONFIG();
-    httpd_handle_t server_handle = NULL;
-    httpd_start(&server_handle, &server_config);
-
-    httpd_uri_t uri_get = {
-        .uri = "/",
-        .method = HTTP_GET,
-        .handler = get_handler,
-        .user_ctx = NULL};
-    httpd_register_uri_handler(server_handle, &uri_get);
-}
 
 void app_main(void)
 {
@@ -37,7 +19,7 @@ void app_main(void)
     ESP_LOGI(TAG, "starting mdns");
     start_mdns_service();
     ESP_LOGI(TAG, "starting http server");
-    start_http_server();
+    // start_http_server();
     ESP_LOGI(TAG, "setting up the led strip");
     strip_setup();
     bool led_on_off = true;
